Added exercise configure service. Updated requests.
This should be local datasource in the client!

+++ src/main/java/com/koleff/kare/exercise/models/request/DeleteExerciseSetRequest.java
@@ -2,5 +2,5 @@ package com.koleff.kare.exercise.models.request;
 
 import java.util.UUID;
 
-public record DeleteExerciseSetRequest(Long exerciseId, Integer workoutId, UUID setId) {
+public record DeleteExerciseSetRequest(Long exerciseId, Long workoutId, UUID setId) {
 }
